+++ MaterialDesignExtensionsDemo/Controls/StepperControl.xaml.cs
@@ -13,6 +13,12 @@ using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
 
+using MaterialDesignThemes.Wpf;
+
+using MaterialDesignExtensions.Controls;
+
+using MaterialDesignExtensionsDemo.ViewModel;
+
 namespace MaterialDesignExtensionsDemo.Controls
 {
     public partial class StepperControl : UserControl
@@ -20,6 +26,16 @@ namespace MaterialDesignExtensionsDemo.Controls
         public StepperControl()
         {
             InitializeComponent();
+
+            stepper.NavigationCanceledByValidation += Stepper_NavigationCanceledByValidation;
+        }
+
+        private void Stepper_NavigationCanceledByValidation(object sender, NavigationCanceledByValidationEventArgs args)
+        {
+            if (args.StepWithValidationErrors.Content is StepperTutorialFourViewModel)
+            {
+                DialogHost.Show(new AlertDialogViewModel("Accept the license terms please."), MainWindow.DialogHostName);
+            }
         }
     }
 }